% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_SGF}
\alias{data_SGF}
\title{Table des données de la SGF}
\format{An object of class \code{data.frame} with 4639686 rows and 9 columns.}
\source{
\url{https://www.insee.fr/fr/statistiques/2591397}
}
\usage{
data_SGF
}
\description{
Données de la SGF (au format long) publiées par l'Insee.
}
\keyword{datasets}
