% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_ARR_SGF_histo}
\alias{geo_ARR_SGF_histo}
\title{Géographies des arrondissements}
\format{Table géographique (format sf data.frame) de 1084 lignes et 3 variables.
\describe{
  \item{CODGEO}{Identifiant du territoire}
  \item{LIBGEO}{Nom du territoire}
  \item{ANNEE_GEOGRAPHIE}{Millésime du contour géographique (1801, 1826 ou 1876)}
  \item{geometry}{Description des contours géographiques}
}}
\usage{
geo_ARR_SGF_histo
}
\description{
Contours et libellés des arrondissements de France métropolitaine à plusieurs millésimes.
}
\keyword{datasets}
