% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_DEP_SGF_histo}
\alias{geo_DEP_SGF_histo}
\title{Géographies des départements}
\format{Table géographique (format sf data.frame) de 437 lignes et 3 variables.
\describe{
  \item{CODGEO}{Identifiant du territoire}
  \item{LIBGEO}{Nom du territoire}
  \item{ANNEE_GEOGRAPHIE}{Millésime du contour géographique (1801, 1826, 1866, 1876 ou 1918)}
  \item{geometry}{Description des contours géographiques}
}}
\usage{
geo_DEP_SGF_histo
}
\description{
Contours et libellés des départements de France métropolitaine à plusieurs millésimes.
}
\keyword{datasets}
